Added a few AvxColumnVector operators: - MultiplyVecVec - SubtractVecVec - OuterProduct (vector * row)
Added some extension methods to convert Avx* to Matrix2D and ColumnVector (the old naive operations).

Converted FeedForward and BackProp to Avx, but kept the older methods. Will rename them later and allow renderer to pick one or the other so we can compare results and benchmark. See RenderContext.cs for changes here.

- Added an Avx OuterProduct test