- Converted most everything to avx* (matrix and column) - Add Layer abstract base class - Added ConvolutionRenderer (in progress), and ConvolutionLayer, and bunch of convolution realted operations. And a test. - Created an abstract base class for the NeuralNetwork.
In trying to create multipe layer types, that take multiple types of inputs and generate different outputs, all the type nonsense got way too convoluted (even though it is possible to make everything generic). I decided to create an intermediary container class that would be the common language between everything (to contain vectors and matrices). This approach greatly simplified the code.

So, I added 'Tensor' which can contain a list of matrices or just a matrix or a vector etc... I think this is the right way to go in order to support different layer types and different inputs/ouputs in the same graph.

Also added a "typeExpriments" project to help think through different approach to accomodate heterogenous graphs and also try to support multiple math implementations (gpu, avx, software).